Start/Pause Menu Updates
StartMenu script commented to describe how to change the scene it loads. Pause Menu doesn't deactivate anymore but can be changed to deactivate if scene is setup with a GameplayObjects empty object. Should be updated once official scene is updated. Also uses timescale=0 which can be dubious depending on how we want the game to work. On the plus side the PauseButton is a working Prefab